refactor(audit): remove dead SIEM transport so the audit module stands alone
The SIEM subsystem carried ~2300 lines of never-wired code: outbound
transports, buffering/flush loops, a duplicate integration service, and
a file-integrity service duplicating apps/audit/tasks.py. None of it had
a production caller. The audit module must work with no SIEM configured.

- Reshape siem.py into a formatting/export service: keep the CEF, LEEF,
  JSON, Syslog, and OCSF renderers, SIEMConfig, and the cache-backed
  HashChainManager; delete SIEMTransport, buffering, flush loops,
  log_to_siem, and the transport-side verify_log_integrity.
- Delete siem_integration.py and file_integrity_service.py outright.
- Drop the unused SIEMExportLog and SIEMHashChainState models
  (+ migration) and their model-allowlist / linter references.
- Rewire the compliance log-integrity report and the audit_compliance
  verify-integrity subcommand onto AuditIntegrityService: a verification
  that cannot run now reports non-compliant (ISO 27001 A.12.4.2
  violation) and the command exits non-zero instead of printing success.
- Remove dead logging classes (RequestIDFilter, ComplianceLogFormatter)
  from audit logging_formatters; the wired filter lives in
  apps/common/logging.py.
- Port format-renderer and hash-chain tests to test_siem_formats.py;
  retire the transport test modules; update docs to the standalone
  contract (structured log stream + on-demand export, no in-app push).

Refs #241
