+++ services/platform/apps/audit/logging_formatters.py
@@ -401,122 +401,3 @@ class AuditLogFormatter(logging.Formatter):
 # =============================================================================
 # REQUEST ID FILTER (MOVED FROM common.logging FOR COMPLETENESS)
 # =============================================================================
-
-
-class RequestIDFilter(logging.Filter):
-    """
-    Add request ID to log records.
-
-    This is a simplified version - the main implementation may be
-    in apps.common.logging. This ensures the filter exists.
-    """
-
-    def filter(self, record: logging.LogRecord) -> bool:
-        """Add request_id attribute to log record"""
-        if not hasattr(record, "request_id"):
-            # Try to get from thread local storage
-            record.request_id = getattr(_audit_context, "request_id", None) or "-" * 36
-        return True
-
-
-# =============================================================================
-# COMPLIANCE LOG FORMATTER
-# =============================================================================
-
-
-class ComplianceLogFormatter(logging.Formatter):
-    """
-    Formatter for compliance-specific logging.
-
-    Produces logs that meet regulatory requirements:
-    - GDPR (EU)
-    - Romanian fiscal code
-    - ISO 27001
-    - SOC 2
-    """
-
-    COMPLIANCE_FRAMEWORKS: ClassVar[dict[str, list[str]]] = {
-        "authentication": ["ISO27001-A.9.4", "SOC2-CC6.1", "GDPR-Art32"],
-        "authorization": ["ISO27001-A.9.2", "SOC2-CC6.2"],
-        "data_modification": ["GDPR-Art30", "ISO27001-A.12.4"],
-        "security_event": ["ISO27001-A.16", "SOC2-CC7.2"],
-        "data_access": ["GDPR-Art15", "ISO27001-A.9.4"],
-        "configuration": ["ISO27001-A.12.1", "SOC2-CC6.6"],
-    }
-
-    def format(self, record: logging.LogRecord) -> str:
-        """Format log for compliance requirements"""
-        timestamp = datetime.utcfromtimestamp(record.created).isoformat() + "Z"
-        category = self._determine_category(record)
-        frameworks = self.COMPLIANCE_FRAMEWORKS.get(category, [])
-
-        compliance_entry = {
-            "@timestamp": timestamp,
-            "compliance": {
-                "frameworks": frameworks,
-                "category": category,
-                "retention_required": True,
-                "sensitive_data": self._contains_sensitive_data(record),
-            },
-            "event": {
-                "action": getattr(record, "action", record.funcName),
-                "outcome": "success" if record.levelno < logging.ERROR else "failure",
-            },
-            "message": record.getMessage(),
-            "log": {
-                "level": record.levelname.lower(),
-                "logger": record.name,
-            },
-            "user": {
-                "id": getattr(record, "user_id", None),
-                "email": getattr(record, "user_email", None),
-            },
-            "source": {
-                "ip": getattr(record, "ip_address", None),
-            },
-            "trace": {
-                "id": getattr(record, "request_id", "-"),
-            },
-        }
-
-        return json.dumps(compliance_entry, default=str, ensure_ascii=False)
-
-    def _determine_category(  # noqa: PLR0911  # Complexity: multi-step business logic
-        self, record: logging.LogRecord
-    ) -> str:  # Complexity: log format selection  # Complexity: multi-step business logic
-        """Determine compliance category"""
-        message = record.getMessage().lower()
-
-        if "login" in message or "logout" in message or "auth" in message:
-            return "authentication"
-        elif "permission" in message or "role" in message or "access" in message:
-            return "authorization"
-        elif "create" in message or "update" in message or "delete" in message:
-            return "data_modification"
-        elif "security" in message or "breach" in message:
-            return "security_event"
-        elif "export" in message or "download" in message:
-            return "data_access"
-        elif "config" in message or "setting" in message:
-            return "configuration"
-
-        return "general"
-
-    def _contains_sensitive_data(self, record: logging.LogRecord) -> bool:
-        """Check if log contains sensitive data indicators"""
-        message = record.getMessage().lower()
-        sensitive_indicators = [
-            "password",
-            "credit",
-            "card",
-            "ssn",
-            "tax",
-            "cui",
-            "bank",
-            "account",
-            "secret",
-            "token",
-            "key",
-            "credential",
-        ]
-        return any(indicator in message for indicator in sensitive_indicators)
