+++ services/platform/tests/audit/test_audit_logging_formatters.py
@@ -21,8 +21,6 @@ from django.test import SimpleTestCase
 from apps.audit.logging_formatters import (
     AuditContextFilter,
     AuditLogFormatter,
-    ComplianceLogFormatter,
-    RequestIDFilter,
     SIEMJSONFormatter,
     clear_audit_context,
     get_audit_context,
@@ -225,44 +223,6 @@ class TestAuditContextFilter(SimpleTestCase):
             self.filter.filter(record)
             # environment should be set (default "production")
             self.assertTrue(hasattr(record, "environment"))
-
-
-# =============================================================================
-# RequestIDFilter
-# =============================================================================
-
-
-class TestRequestIDFilter(SimpleTestCase):
-    """Tests for RequestIDFilter.filter()."""
-
-    def setUp(self) -> None:
-        self.filter = RequestIDFilter()
-
-    def tearDown(self) -> None:
-        clear_audit_context()
-
-    def test_filter_always_returns_true(self) -> None:
-        record = make_record()
-        self.assertTrue(self.filter.filter(record))
-
-    def test_adds_request_id_when_absent(self) -> None:
-        set_audit_context(request_id="req-from-context")
-        record = make_record()
-        # Ensure record does NOT already have request_id
-        self.assertFalse(hasattr(record, "request_id"))
-        self.filter.filter(record)
-        self.assertEqual(record.request_id, "req-from-context")
-
-    def test_does_not_overwrite_existing_request_id(self) -> None:
-        set_audit_context(request_id="context-id")
-        record = make_record(request_id="existing-id")
-        self.filter.filter(record)
-        self.assertEqual(record.request_id, "existing-id")
-
-    def test_request_id_dash_when_context_empty(self) -> None:
-        record = make_record()
-        self.filter.filter(record)
-        self.assertEqual(record.request_id, "-" * 36)
 
 
 # =============================================================================
@@ -610,220 +570,3 @@ class TestAuditLogFormatterDetermineCategory(SimpleTestCase):
 
     def test_default_category(self) -> None:
         self.assertEqual(self._category("something unrelated happened"), "general")
-
-
-# =============================================================================
-# ComplianceLogFormatter
-# =============================================================================
-
-
-class TestComplianceLogFormatterFormat(SimpleTestCase):
-    """Tests for ComplianceLogFormatter.format()."""
-
-    def setUp(self) -> None:
-        self.formatter = ComplianceLogFormatter()
-
-    def _format_and_parse(self, record: logging.LogRecord) -> dict[str, Any]:
-        output = self.formatter.format(record)
-        return json.loads(output)
-
-    def test_output_is_valid_json(self) -> None:
-        record = make_record()
-        output = self.formatter.format(record)
-        self.assertIsInstance(json.loads(output), dict)
-
-    def test_required_top_level_keys(self) -> None:
-        record = make_record()
-        data = self._format_and_parse(record)
-        for key in ("@timestamp", "compliance", "event", "message", "log", "user", "source", "trace"):
-            self.assertIn(key, data, f"Missing key: {key}")
-
-    def test_compliance_block_structure(self) -> None:
-        record = make_record()
-        data = self._format_and_parse(record)
-        comp = data["compliance"]
-        self.assertIn("frameworks", comp)
-        self.assertIn("category", comp)
-        self.assertIn("retention_required", comp)
-        self.assertIn("sensitive_data", comp)
-        self.assertTrue(comp["retention_required"])
-
-    def test_timestamp_ends_with_z(self) -> None:
-        record = make_record()
-        data = self._format_and_parse(record)
-        self.assertTrue(data["@timestamp"].endswith("Z"))
-
-    def test_authentication_frameworks(self) -> None:
-        record = make_record(msg="User login attempt")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("ISO27001-A.9.4", frameworks)
-        self.assertIn("SOC2-CC6.1", frameworks)
-        self.assertIn("GDPR-Art32", frameworks)
-
-    def test_authorization_frameworks(self) -> None:
-        record = make_record(msg="Permission check for role")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("ISO27001-A.9.2", frameworks)
-        self.assertIn("SOC2-CC6.2", frameworks)
-
-    def test_data_modification_frameworks(self) -> None:
-        record = make_record(msg="create new record")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("GDPR-Art30", frameworks)
-        self.assertIn("ISO27001-A.12.4", frameworks)
-
-    def test_security_event_frameworks(self) -> None:
-        record = make_record(msg="security breach detected")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("ISO27001-A.16", frameworks)
-        self.assertIn("SOC2-CC7.2", frameworks)
-
-    def test_data_access_frameworks(self) -> None:
-        record = make_record(msg="data export completed")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("GDPR-Art15", frameworks)
-
-    def test_configuration_frameworks(self) -> None:
-        # Avoid "update" which matches data_modification first
-        record = make_record(msg="config saved for module")
-        data = self._format_and_parse(record)
-        frameworks = data["compliance"]["frameworks"]
-        self.assertIn("ISO27001-A.12.1", frameworks)
-        self.assertIn("SOC2-CC6.6", frameworks)
-
-    def test_general_category_empty_frameworks(self) -> None:
-        record = make_record(msg="something happened")
-        data = self._format_and_parse(record)
-        self.assertEqual(data["compliance"]["frameworks"], [])
-
-    def test_outcome_success_for_non_error(self) -> None:
-        record = make_record(level=logging.INFO)
-        data = self._format_and_parse(record)
-        self.assertEqual(data["event"]["outcome"], "success")
-
-    def test_outcome_failure_for_error(self) -> None:
-        record = make_record(level=logging.ERROR)
-        data = self._format_and_parse(record)
-        self.assertEqual(data["event"]["outcome"], "failure")
-
-
-class TestComplianceLogFormatterDetermineCategory(SimpleTestCase):
-    """Tests for ComplianceLogFormatter._determine_category()."""
-
-    def setUp(self) -> None:
-        self.formatter = ComplianceLogFormatter()
-
-    def _category(self, msg: str) -> str:
-        return self.formatter._determine_category(make_record(msg=msg))
-
-    def test_login(self) -> None:
-        self.assertEqual(self._category("login attempt"), "authentication")
-
-    def test_logout(self) -> None:
-        self.assertEqual(self._category("logout event"), "authentication")
-
-    def test_auth(self) -> None:
-        self.assertEqual(self._category("auth failed"), "authentication")
-
-    def test_permission(self) -> None:
-        self.assertEqual(self._category("permission denied"), "authorization")
-
-    def test_role(self) -> None:
-        self.assertEqual(self._category("role changed"), "authorization")
-
-    def test_access(self) -> None:
-        self.assertEqual(self._category("access revoked"), "authorization")
-
-    def test_create(self) -> None:
-        self.assertEqual(self._category("create entity"), "data_modification")
-
-    def test_update(self) -> None:
-        self.assertEqual(self._category("update entity"), "data_modification")
-
-    def test_delete(self) -> None:
-        self.assertEqual(self._category("delete entity"), "data_modification")
-
-    def test_security(self) -> None:
-        self.assertEqual(self._category("security event"), "security_event")
-
-    def test_breach(self) -> None:
-        self.assertEqual(self._category("breach occurred"), "security_event")
-
-    def test_export(self) -> None:
-        self.assertEqual(self._category("export data"), "data_access")
-
-    def test_download(self) -> None:
-        self.assertEqual(self._category("download report"), "data_access")
-
-    def test_config(self) -> None:
-        self.assertEqual(self._category("config saved"), "configuration")
-
-    def test_setting(self) -> None:
-        self.assertEqual(self._category("setting changed"), "configuration")
-
-    def test_general_default(self) -> None:
-        self.assertEqual(self._category("routine check"), "general")
-
-
-class TestComplianceLogFormatterContainsSensitiveData(SimpleTestCase):
-    """Tests for ComplianceLogFormatter._contains_sensitive_data()."""
-
-    def setUp(self) -> None:
-        self.formatter = ComplianceLogFormatter()
-
-    def _sensitive(self, msg: str) -> bool:
-        return self.formatter._contains_sensitive_data(make_record(msg=msg))
-
-    def test_password_indicator(self) -> None:
-        self.assertTrue(self._sensitive("Reset password for user"))
-
-    def test_credit_indicator(self) -> None:
-        self.assertTrue(self._sensitive("credit card number"))
-
-    def test_card_indicator(self) -> None:
-        self.assertTrue(self._sensitive("card details updated"))
-
-    def test_ssn_indicator(self) -> None:
-        self.assertTrue(self._sensitive("SSN provided"))
-
-    def test_tax_indicator(self) -> None:
-        self.assertTrue(self._sensitive("tax ID validation"))
-
-    def test_cui_indicator(self) -> None:
-        self.assertTrue(self._sensitive("CUI lookup for RO12345678"))
-
-    def test_bank_indicator(self) -> None:
-        self.assertTrue(self._sensitive("bank details saved"))
-
-    def test_account_indicator(self) -> None:
-        self.assertTrue(self._sensitive("account number changed"))
-
-    def test_secret_indicator(self) -> None:
-        self.assertTrue(self._sensitive("secret key rotated"))
-
-    def test_token_indicator(self) -> None:
-        self.assertTrue(self._sensitive("token generated"))
-
-    def test_key_indicator(self) -> None:
-        self.assertTrue(self._sensitive("api key issued"))
-
-    def test_credential_indicator(self) -> None:
-        self.assertTrue(self._sensitive("credential stored"))
-
-    def test_normal_message_not_sensitive(self) -> None:
-        self.assertFalse(self._sensitive("User login successful"))
-
-    def test_empty_message_not_sensitive(self) -> None:
-        self.assertFalse(self._sensitive(""))
-
-    def test_unrelated_message_not_sensitive(self) -> None:
-        self.assertFalse(self._sensitive("Invoice generated for customer"))
-
-    def test_case_insensitive_detection(self) -> None:
-        self.assertTrue(self._sensitive("PASSWORD must be changed"))
-        self.assertTrue(self._sensitive("User TOKEN expired"))
